RavenDB-21064 - addressing PR comments
