+++ src/Raven.Server/Commercial/LicenseAttribute.cs
@@ -20,8 +20,8 @@ public enum LicenseAttribute
     RavenEtl,
     SqlEtl,
     HighlyAvailableTasks,
-    PullReplicationAsHub,
-    PullReplicationAsSink,
+    PullReplicationHub,
+    PullReplicationSink,
     EncryptedBackup,
     LetsEncryptAutoRenewal,
     Cloud,
