+++ src/Raven.Server/Documents/Handlers/Processors/Indexes/AbstractIndexHandlerProcessorForSuggestIndexMerge.cs
@@ -15,7 +15,7 @@ internal abstract class AbstractIndexHandlerProcessorForSuggestIndexMerge<TReque
 {
     protected AbstractIndexHandlerProcessorForSuggestIndexMerge([NotNull] TRequestHandler requestHandler) : base(requestHandler)
     {
-        if (requestHandler.ServerStore.LicenseManager.LicenseStatus.HasIndexCleanup)
+        if (requestHandler.ServerStore.LicenseManager.LicenseStatus.HasIndexCleanup == false)
         {
             throw new LicenseLimitException(LimitType.IndexCleanup, "Your license doesn't include the index suggestions feature.");
         }
