+++ src/Raven.Server/Documents/Handlers/Processors/OngoingTasks/AbstractOngoingTasksHandlerProcessorForUpdatePeriodicBackup.cs
@@ -2,7 +2,6 @@
 using JetBrains.Annotations;
 using Microsoft.AspNetCore.Http.Features.Authentication;
 using Raven.Client.Documents.Operations.Backups;
-using Raven.Client.Exceptions.Commercial;
 using Raven.Server.Documents.Handlers.Processors.Databases;
 using Raven.Server.Documents.PeriodicBackup;
 using Raven.Server.ServerWide;
